рефакторит модуль с вспомогательными ф-имя
- удалит utils.js
- ф-ии из utils.js перенесёт в 2 файла (common.js и point.js)
