+++ src/utils/common.js
@@ -1,5 +1,3 @@
-import dayjs from 'dayjs';
-
 const getRandomInteger = (a, b) => {
   const lower = Math.ceil(Math.min(a, b));
   const upper = Math.floor(Math.max(a, b));
@@ -45,49 +43,15 @@ const createRandomNumberFromRange = (
   };
 };
 
-const getDateDiff = (fromDate, toDate, unit = 'ms') =>
-  dayjs(toDate).diff(dayjs(fromDate), unit);
-
-const formatDate = (date, typeFormating) =>
-  date ? dayjs(date).format(typeFormating) : '';
-
-const getDate = ({ next }) => {
-  const currentDate = new Date();
-
-  return next
-    ? dayjs(currentDate).add(getRandomInteger(0, 60), 'minute').add(getRandomInteger(0, 24), 'hour').add(getRandomInteger(0, 28), 'day').toDate() : dayjs().toDate();
-};
-
-const getFormattedDateDifference = (dateFrom, dateTo) => {
-  const dayCount = getDateDiff(dateFrom, dateTo, 'd');
-  const hourCount = getDateDiff(dateFrom, dateTo, 'h');
-  const minuteCount = getDateDiff(dateFrom, dateTo, 'm');
-  let formattedDate;
-
-  if (dayCount === 0 && hourCount === 0) {
-    formattedDate = `${minuteCount}M`;
-  } else if (dayCount === 0) {
-    formattedDate = `${hourCount}H ${minuteCount - hourCount * 60}M`;
-  } else {
-    formattedDate = `${dayCount}D ${hourCount - dayCount * 24}H ${
-      minuteCount - hourCount * 60
-    }M`;
-  }
-  return formattedDate;
-};
-
 const capitalizeFirstLetter = (string) =>
   string.charAt(0).toUpperCase() + string.slice(1);
 
 const findObject = (arr, key, value) => arr.find((obj) => obj[key] === value);
 
 export {
-  createRandomNumberFromRange,
+  getRandomInteger,
   getRandomArrayElement,
-  formatDate,
-  getDateDiff,
-  getFormattedDateDifference,
-  getDate,
+  createRandomNumberFromRange,
   capitalizeFirstLetter,
   findObject,
 };
